# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Danshin Genealogy",
        "description": "Доступ к большенству данных только через API токен,\n        в противном случае будет возвращаться ошибка 401.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/dates": {
            "get": {
                "description": "Список всех известных дат (день рождения, день памяти). Под известными понимаются даты, с полными данными о дне, месяце, годе. Например дата 2000-11-1? не попадёт, из-за неизевстного дня.",
                "operationId": "fdfd1a1676f59139ddc326a711ca3bef",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/datesDate"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация через токен"
                    }
                }
            }
        },
        "/api/dates/upcoming": {
            "get": {
                "description": "Список ближайших дат (предыдущие дни, текущий день, будущие дни, в соответтвии с переданными параметрами).",
                "operationId": "4bbd43443fba95f7662a6111cbb24add",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Текущее число в формате гггг-мм-дд",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "past_day",
                        "in": "query",
                        "description": "Количество предшествующих дней (число от 1 до 30).",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nearest_day",
                        "in": "query",
                        "description": "Количество будующих дней (число от 1 до 30).",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/datesEvents"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация через токен"
                    },
                    "422": {
                        "description": "Неверные параметры запроса"
                    }
                }
            }
        },
...
        "/api/person-photo/{personId}": {
            "get": {
                "description": "Массив описаний фотоснимков для лица (!!!не файлы).",
                "operationId": "43bfe400bc66e3290f33b09279d855bb",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "Id лица.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/personReaderPhoto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация через токен"
                    },
                    "404": {
                        "description": "Лица с переданным id не найдено."
                    }
                }
            }
        },
        "/api/photo/{personId}/{fileName}": {
            "get": {
                "description": "Файл фотографии (бинарный).",
                "operationId": "81c977c2583795ca1aa193b9b186968e",
                "parameters": [
                    {
                        "name": "personId",
                        "in": "path",
                        "description": "Id лица.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "fileName",
                        "in": "path",
                        "description": "Имя файла фото.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "image/jpg": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация через токен"
                    },
                    "404": {
                        "description": "Фото с переданными параметрами не найдено."
                    }
                }
            }
        },
        "/api/photo/temp": {
            "post": {
                "description": "Сохранить файл фотоснимка во временное хранилище",
                "operationId": "989058617c3b1d1cfd4323959b4933ab",
                "responses": {
                    "200": {
                        "description": "OK. ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "": {
                                            "description": "Имя файла в поле fileName",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Требуется авторизация через токен"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "dateInterval": {
                "required": [
                    "y",
                    "m",
                    "d"
                ],
                "properties": {
                    "y": {
                        "description": "Год",
                        "type": "integer"
                    },
                    "m": {
                        "description": "Месяц",
                        "type": "integer"
                    },
                    "d": {
                        "description": "День",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "date": {
                "title": "Дата.",
                "required": [
                    "hasUnknown",
                    "isEmpty"
                ],
                "properties": {
                    "year": {
                        "description": "Год.",
                        "type": "string",
                        "nullable": true
                    },
                    "month": {
                        "description": "Месяц.",
                        "type": "string",
                        "nullable": true
                    },
                    "day": {
                        "description": "День.",
                        "type": "string",
                        "nullable": true
                    },
                    "hasUnknown": {
                        "description": "Является ли дата неизестной. Дата неизвестна, если год или месяц или день, имеют хотя бы одну известную цифру. Неизвестные данные будут заменены на ? (вопросительный знак). ",
                        "type": "boolean"
                    },
                    "isEmpty": {
                        "description": "Является ли дата пустой. Дата пустая, если год и месяц и день пустые (проверка каждого элемента на empty() == true).",
                        "type": "boolean"
                    },
                    "string": {
                        "description": "Строка даты в формате YYYY-MM-DD.",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "datesBirth": {
                "title": "День рождения.",
                "required": [
                    "date",
                    "person",
                    "age"
                ],
                "properties": {
                    "date": {
                        "$ref": "#/components/schemas/date"
                    },
                    "person": {
                        "$ref": "#/components/schemas/datesPerson"
                    },
                    "age": {
                        "$ref": "#/components/schemas/dateInterval"
                    }
                },
                "type": "object"
            },
            "datesBirthWould": {
                "title": "День рождения умершего лица.",
                "required": [
                    "date",
                    "person",
                    "age"
                ],
                "properties": {
                    "date": {
                        "$ref": "#/components/schemas/date"
                    },
                    "person": {
                        "$ref": "#/components/schemas/datesPerson"
                    },
                    "age": {
                        "$ref": "#/components/schemas/dateInterval"
                    }
                },
                "type": "object"
            },
            "datesDate": {
                "title": "Дата события.",
                "required": [
                    "date",
                    "type",
                    "person"
                ],
                "properties": {
                    "date": {
                        "$ref": "#/components/schemas/date"
                    },
                    "type": {
                        "description": "Тип даты.",
                        "type": "App\\Models\\Dates\\DateType",
                        "enum": [
                            1,
                            2
                        ]
                    },
                    "person": {
                        "$ref": "#/components/schemas/datesPerson"
                    }
                },
                "type": "object"
            },
            "datesDeath": {
                "title": "День смерти.",
                "required": [
                    "date",
                    "person",
                    "interval"
                ],
                "properties": {
                    "date": {
                        "$ref": "#/components/schemas/date"
                    },
                    "person": {
                        "$ref": "#/components/schemas/datesPerson"
                    },
                    "age": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/dateInterval"
                            }
                        ],
                        "nullable": true,
                        "description": "Возраст (объект DateInterval).",
                        "type": "DateInterval"
                    },
                    "interval": {
                        "$ref": "#/components/schemas/dateInterval"
                    }
                },
                "type": "object"
            },
            "datesEvents": {
                "title": "Список ближайших событий.",
                "required": [
                    "pastBirth",
                    "pastBirthWould",
                    "pastDeath",
                    "todayBirth",
                    "todayBirthWould",
                    "todayDeath",
                    "nearestBirth",
                    "nearestBirthWould",
                    "nearestDeath",
                    "isEmpty"
                ],
                "properties": {
                    "pastBirth": {
                        "description": "Прошедшие дни рождения живых лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirth"
                        }
                    },
                    "pastBirthWould": {
                        "description": "Прошедшие дни рождения умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirthWould"
                        }
                    },
                    "pastDeath": {
                        "description": "Прошедшие дни памяти умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesDeath"
                        }
                    },
                    "todayBirth": {
                        "description": "Текущие дни рождения живых лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirth"
                        }
                    },
                    "todayBirthWould": {
                        "description": "Текущие дни рождения умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirthWould"
                        }
                    },
                    "todayDeath": {
                        "description": "Текущие дни памяти умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesDeath"
                        }
                    },
                    "nearestBirth": {
                        "description": "Будующие дни рождения живых лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirth"
                        }
                    },
                    "nearestBirthWould": {
                        "description": "Будующие дни рождения умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesBirthWould"
                        }
                    },
                    "nearestDeath": {
                        "description": "Будущие дни памяти умерших лиц.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/datesDeath"
                        }
                    },
                    "isEmpty": {
                        "description": "Отсутствуют ли события.",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "datesPerson": {
                "title": "Лицо, связанное с событием.",
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "Id",
                        "type": "integer"
                    },
                    "surname": {
                        "description": "Фамилия",
                        "type": "string",
                        "nullable": true
                    },
                    "oldSurname": {
                        "description": "Прежнии фамилии",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "nullable": true
                    },
                    "name": {
                        "description": "Имя",
                        "type": "string",
                        "nullable": true
                    },
                    "patronymic": {
                        "description": "Отчество",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "personReaderPhoto": {
                "title": "Описание фотоснимка",
                "required": [
                    "order",
                    "fileName"
                ],
                "properties": {
                    "order": {
                        "description": "Порядковый номер фотоснимка.",
                        "type": "integer"
                    },
                    "fileName": {
                        "description": "Имя файла.",
                        "type": "string"
                    },
                    "date": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/date"
                            }
                        ],
                        "nullable": true,
                        "description": "Дата."
                    }
                },
                "type": "object"
            },
...
        }
    }
}